+++ src/commands.rs
@@ -13,7 +13,7 @@ use crate::verify::{self, Algorithm};
 
 #[derive(Debug)]
 pub struct CommandResult {
-    pub file_location: PathBuf,
+    pub file_location: Option<PathBuf>,
     pub used_algorithm: Algorithm,
     pub calculated_hash_sum: String,
     pub hash_compare_result: Option<HashCompareResult>,
@@ -43,7 +43,7 @@ impl fmt::Display for CommandError {
                 write!(f, "{}", msg)
             }
             CommandError::InvalidUrl => {
-                write!(f, "The specified url is invalid. An example of a valid URL is 'https://testfile.org'")
+                write!(f, "The provided URL is invalid. Please ensure the URL is correctly formatted,\nincluding the scheme (e.g. 'http://', 'https://').\nFor example: https://www.example.com")
             }
             CommandError::OutputTargetInvalid(target) => {
                 let msg = format!(
@@ -63,6 +63,7 @@ impl fmt::Display for CommandError {
     }
 }
 
+// Handle the CLI subcommand 'download'
 pub fn handle_download_cmd(args: DownloadArgs, os_type: os_specifics::OS) -> Result<()> {
     // fetch the output target
     let output_target = args.output;
@@ -116,10 +117,13 @@ pub fn handle_download_cmd(args: DownloadArgs, os_type: os_specifics::OS) -> Res
     };
 
     // start the download
-    let file_location = download::execute_download(download_properties)?;
+    let file_path = download::execute_download(download_properties)?;
+
+    // Reads the entire contents of the downloaded file into a bytes vector.
+    let data_to_hash = std::fs::read(&file_path)?;
 
     let mut calculated_hash_sum =
-        verify::get_hash_sum_as_lower_hex(&file_location, args.algorithm)?;
+        verify::get_hash_sum_as_lower_hex(data_to_hash, Some(file_path.clone()), args.algorithm)?;
 
     let cmd_result = if let Some(origin_hash_sum) = args.hash_sum {
         if !verify::is_lower_hex(&origin_hash_sum) {
@@ -132,7 +136,7 @@ pub fn handle_download_cmd(args: DownloadArgs, os_type: os_specifics::OS) -> Res
 
         let is_file_modified = verify::is_hash_equal(&origin_hash_sum, &calculated_hash_sum);
         CommandResult {
-            file_location,
+            file_location: Some(file_path),
             used_algorithm: args.algorithm,
             calculated_hash_sum,
             hash_compare_result: Some(HashCompareResult {
@@ -142,7 +146,7 @@ pub fn handle_download_cmd(args: DownloadArgs, os_type: os_specifics::OS) -> Res
         }
     } else {
         CommandResult {
-            file_location,
+            file_location: Some(file_path),
             used_algorithm: args.algorithm,
             calculated_hash_sum,
             hash_compare_result: None,
@@ -152,44 +156,56 @@ pub fn handle_download_cmd(args: DownloadArgs, os_type: os_specifics::OS) -> Res
     Ok(())
 }
 
+// Handle the CLI subcommand 'local'
 pub fn handle_local_cmd(args: LocalArgs) -> Result<()> {
-    let source_file = Path::new(&args.file_path);
-    // check if the given file exist
-    if source_file.exists() {
-        let mut calculated_hash_sum =
-            verify::get_hash_sum_as_lower_hex(source_file, args.algorithm)?;
-
-        let cmd_result = if let Some(origin_hash_sum) = args.hash_sum {
-            if !verify::is_lower_hex(&origin_hash_sum) {
-                // convert the calculated hash sum to UpperHex
-                calculated_hash_sum = calculated_hash_sum
-                    .chars()
-                    .map(|c| c.to_uppercase().to_string())
-                    .collect();
-            }
-
-            let is_file_modified = verify::is_hash_equal(&origin_hash_sum, &calculated_hash_sum);
-            CommandResult {
-                file_location: source_file.to_path_buf(),
-                used_algorithm: args.algorithm,
-                calculated_hash_sum,
-                hash_compare_result: Some(HashCompareResult {
-                    is_file_modified,
-                    origin_hash_sum,
-                }),
-            }
+    let (data_to_hash, file_path) = if let Some(file) = args.file {
+        // check if the given file path point to an existing file
+        let file_path = PathBuf::from(file);
+        if file_path.exists() {
+            let data_to_hash = std::fs::read(file_path.clone())?;
+            (data_to_hash, Some(file_path))
         } else {
-            CommandResult {
-                file_location: source_file.to_path_buf(),
-                used_algorithm: args.algorithm,
-                calculated_hash_sum,
-                hash_compare_result: None,
-            }
-        };
-        utils::processing_cmd_result(cmd_result);
-        Ok(())
+            let command_err = CommandError::FileNotExist(utils::get_absolute_path(&file_path));
+            return Err(color_eyre::eyre::eyre!(command_err.to_string()));
+        }
+    } else if let Some(some_text) = args.buffer {
+        let data_to_hash = some_text.as_bytes().to_vec();
+        (data_to_hash, None)
     } else {
-        let command_err = CommandError::FileNotExist(utils::get_absolute_path(source_file));
-        Err(color_eyre::eyre::eyre!(command_err.to_string()))
-    }
+        return Ok(());
+    };
+
+    let mut calculated_hash_sum =
+        verify::get_hash_sum_as_lower_hex(data_to_hash, file_path.clone(), args.algorithm)?;
+
+    let cmd_result = if let Some(origin_hash_sum) = args.hash_sum {
+        if !verify::is_lower_hex(&origin_hash_sum) {
+            // convert the calculated hash sum to UpperHex
+            calculated_hash_sum = calculated_hash_sum
+                .chars()
+                .map(|c| c.to_uppercase().to_string())
+                .collect();
+        }
+
+        let is_file_modified = verify::is_hash_equal(&origin_hash_sum, &calculated_hash_sum);
+        CommandResult {
+            file_location: file_path,
+            used_algorithm: args.algorithm,
+            calculated_hash_sum,
+            hash_compare_result: Some(HashCompareResult {
+                is_file_modified,
+                origin_hash_sum,
+            }),
+        }
+    } else {
+        CommandResult {
+            file_location: file_path,
+            used_algorithm: args.algorithm,
+            calculated_hash_sum,
+            hash_compare_result: None,
+        }
+    };
+    utils::processing_cmd_result(cmd_result);
+
+    Ok(())
 }